+++ cmd/rpcdaemon/main.go
@@ -7,13 +7,11 @@ import (
 	"github.com/ledgerwatch/erigon/cmd/rpcdaemon/commands"
 	"github.com/ledgerwatch/erigon/cmd/rpcdaemon/filters"
 	"github.com/ledgerwatch/erigon/cmd/utils"
-	"github.com/ledgerwatch/erigon/common/fdlimit"
 	"github.com/ledgerwatch/log/v3"
 	"github.com/spf13/cobra"
 )
 
 func main() {
-	raiseFdLimit()
 	cmd, cfg := cli.RootCommand()
 	rootCtx, rootCancel := utils.RootContext()
 	cmd.RunE = func(cmd *cobra.Command, args []string) error {
@@ -42,17 +40,5 @@ func main() {
 	if err := cmd.ExecuteContext(rootCtx); err != nil {
 		log.Error(err.Error())
 		os.Exit(1)
-	}
-}
-
-// raiseFdLimit raises out the number of allowed file handles per process
-func raiseFdLimit() {
-	limit, err := fdlimit.Maximum()
-	if err != nil {
-		log.Error("Failed to retrieve file descriptor allowance", "error", err)
-		return
-	}
-	if _, err = fdlimit.Raise(uint64(limit)); err != nil {
-		log.Error("Failed to raise file descriptor allowance", "error", err)
 	}
 }
